requirements.txt
yfinance>=0.2.55
pandas>=2.2.3
numpy>=2.1.3
matplotlib>=3.10.1
requests>=2.32.3
cachetools>=5.5.2
python-dotenv>=1.1.0
ta-lib>=0.6.3
xgboost>=3.0.0
shap>=0.47.1
backtrader>=1.9.78
pyportfolioopt>=1.5.6
fastapi>=0.115.12
uvicorn>=0.34.0
aiohttp>=3.9.3
python-multipart>=0.0.7